jsdoc and some type tests for apps.* API, but manifest APIs needs some thinking. do we want to type them fully?
